==> notebooks/04_topk_recommendations.ipynb <==
import pandas as pd
import pickle

# Load processed data
train_df = pd.read_csv("../data/processed/train_ratings.csv")
movies = pd.read_csv("../data/raw/movies.csv")

# Load trained SVD model
with open("../models/svd_model.pkl", "rb") as f:
    svd = pickle.load(f)

print("Model and data loaded ")


def get_unseen_movies(user_id, train_df, movies_df):
    seen_movies = train_df[train_df["userId"] == user_id]["movieId"].unique()
    all_movies = movies_df["movieId"].unique()
    
    unseen_movies = list(set(all_movies) - set(seen_movies))
    return unseen_movies


def recommend_top_k(user_id, k=10):
    unseen_movies = get_unseen_movies(user_id, train_df, movies)
    
    predictions = []
    for movie_id in unseen_movies:
        pred = svd.predict(user_id, movie_id)
        predictions.append((movie_id, pred.est))
    
    predictions.sort(key=lambda x: x[1], reverse=True)
    top_k = predictions[:k]
    
    return top_k


user_id = 10  # you can change this
top_10 = recommend_top_k(user_id, k=10)

top_10


def show_recommendations(top_k, movies_df):
    result = []
    for movie_id, score in top_k:
        title = movies_df[movies_df["movieId"] == movie_id]["title"].values[0]
        result.append((title, round(score, 3)))
    return result

show_recommendations(top_10, movies)


Top-K Recommendation:
For a given user, unseen movies were ranked based on predicted ratings from the SVD model, and the top-10 highest scoring items were recommended.
